align RecipeCard
